+++ src/main/java/demo.java
@@ -5,11 +5,10 @@ import java.util.List;
 import be.baur.sda.DataNode;
 import be.baur.sda.Node;
 import be.baur.sda.SDA;
-import be.baur.sda.serialization.ParseException;
 
 public class demo {
 
-	public static void main(String[] args) throws IOException, ParseException  {
+	public static void main(String[] args) throws IOException {
 		
 		FileReader fin = new FileReader(args[0]);
 		Node root = SDA.parser().parse(fin);