+++ src/test/java/test/TestSDAParserAlt.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.io.StringReader;
 
 import be.baur.sda.Node;
-import be.baur.sda.serialization.SDASyntaxException;
 import samples.parser.SDAParser;
 
 public final class TestSDAParserAlt {
@@ -69,7 +68,7 @@ public final class TestSDAParserAlt {
 		UnitTestPerformance<String> p = new UnitTestPerformance<String>(str -> {
 			try {
 				parser.parse(new StringReader(str));
-			} catch (SDASyntaxException | IOException e) {
+			} catch (IOException e) {
 				e.printStackTrace();
 			}
 		});